Remove duplicated `Fastlane::Helper` types
Apart from the fact that it's confusing and redundant to have them, and
that they've been a stone in our foot for a while, they result in
runtime issues in Ruby 3.x.
