+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/metadata_update_helper.rb
@@ -50,84 +50,20 @@ module Fastlane
 
         if line_count <= 1
           # Single line output
-          fw.puts("msgid \"#{File.read(@content_file_path)}\"")
+          fw.puts("msgid \"#{File.read(@content_file_path).rstrip}\"")
         else
           # Multiple line output
           fw.puts('msgid ""')
 
           # insert content
-          sf = File.open(@content_file_path, 'r').to_a
-          sf.each do |line|
-            l = "\"#{line.strip}"
-            l << '\\n' unless line == sf.last
-            l << '"'
-            fw.puts(l)
+          File.open(@content_file_path, 'r').each do |line|
+            fw.puts("\"#{line.strip}\\n\"")
           end
         end
 
         # close
         fw.puts('msgstr ""')
         fw.puts('')
-      end
-    end
-
-    class ReleaseNoteMetadataBlock < StandardMetadataBlock
-      attr_reader :new_key, :keep_key, :rel_note_key, :release_version
-
-      def initialize(block_key, content_file_path, release_version)
-        super(block_key, content_file_path)
-        @rel_note_key = 'release_note'
-        @release_version = release_version
-        generate_keys(release_version)
-      end
-
-      def generate_keys(release_version)
-        values = release_version.split('.')
-        version_major = Integer(values[0])
-        version_minor = Integer(values[1])
-        @new_key = "#{@rel_note_key}_#{version_major}#{version_minor}"
-
-        version_major -= 1 if version_minor == 0
-        version_minor = version_minor == 0 ? 9 : version_minor - 1
-
-        @keep_key = "#{@rel_note_key}_#{version_major}#{version_minor}"
-      end
-
-      def is_handler_for(key)
-        values = key.split('_')
-        key.start_with?(@rel_note_key) && values.length == 3 && (!Integer(values[2]).nil? rescue false)
-      end
-
-      def handle_line(fw, line)
-        # put content on block start or if copying the latest one
-        # and skip all the other content
-        if line.start_with?('msgctxt')
-          key = extract_key(line)
-          @is_copying = (key == @keep_key)
-          generate_block(fw) if @is_copying
-        end
-
-        fw.puts(line) if @is_copying
-      end
-
-      def generate_block(fw)
-        # init
-        fw.puts("msgctxt \"#{@new_key}\"")
-        fw.puts('msgid ""')
-        fw.puts("\"#{@release_version}:\\n\"")
-
-        # insert content
-        File.open(@content_file_path, 'r').each do |line|
-          fw.puts("\"#{line.strip}\\n\"")
-        end
-
-        # close
-        fw.puts('msgstr ""')
-        fw.puts('')
-      end
-
-      def extract_key(line)
-        line.split[1].tr('\"', '')
       end
     end
 
@@ -176,6 +112,88 @@ module Fastlane
         # close
         fw.puts('msgstr ""')
         fw.puts('')
+      end
+    end
+
+    class ReleaseNoteMetadataBlock < StandardMetadataBlock
+      attr_reader :new_key, :keep_key, :rel_note_key, :release_version
+
+      def initialize(block_key, content_file_path, release_version)
+        super(block_key, content_file_path)
+        @rel_note_key = 'release_note'
+        @release_version = release_version
+        generate_keys(release_version)
+      end
+
+      def generate_keys(release_version)
+        values = release_version.split('.')
+        version_major = Integer(values[0])
+        version_minor = Integer(values[1])
+        @new_key = "#{@rel_note_key}_#{version_major.to_s.rjust(2, '0')}#{version_minor}"
+
+        version_major -= 1 if version_minor == 0
+        version_minor = version_minor == 0 ? 9 : version_minor - 1
+
+        @keep_key = "#{@rel_note_key}_#{version_major.to_s.rjust(2, '0')}#{version_minor}"
+      end
+
+      def is_handler_for(key)
+        values = key.split('_')
+        key.start_with?(@rel_note_key) && values.length == 3 && is_int?(values[2].sub(/^0*/, ''))
+      end
+
+      def handle_line(fw, line)
+        # put content on block start or if copying the latest one
+        # and skip all the other content
+        if line.start_with?('msgctxt')
+          key = extract_key(line)
+          @is_copying = (key == @keep_key)
+          generate_block(fw) if @is_copying
+        end
+
+        fw.puts(line) if @is_copying
+      end
+
+      def generate_block(fw)
+        # init
+        fw.puts("msgctxt \"#{@new_key}\"")
+        fw.puts('msgid ""')
+        fw.puts("\"#{@release_version}:\\n\"")
+
+        # insert content
+        File.open(@content_file_path, 'r').each do |line|
+          fw.puts("\"#{line.strip}\\n\"")
+        end
+
+        # close
+        fw.puts('msgstr ""')
+        fw.puts('')
+      end
+
+      def extract_key(line)
+        line.split[1].tr('\"', '')
+      end
+
+      def is_int?(value)
+        true if Integer(value) rescue false
+      end
+    end
+
+    class ReleaseNoteShortMetadataBlock < ReleaseNoteMetadataBlock
+      def initialize(block_key, content_file_path, release_version)
+        super(block_key, content_file_path, release_version)
+        @rel_note_key = 'release_note_short'
+        @release_version = release_version
+        generate_keys(release_version)
+      end
+
+      def is_handler_for(key)
+        values = key.split('_')
+        key.start_with?(@rel_note_key) && values.length == 4 && is_int?(values[3].sub(/^0*/, ''))
+      end
+
+      def generate_block(fw)
+        super(fw) unless File.empty?(@content_file_path)
       end
     end
   end
